% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{agentic_sessions}
\alias{agentic_sessions}
\title{List available sessions for resuming}
\usage{
agentic_sessions()
}
\description{
List available sessions for resuming
}
\examples{
\dontrun{
agentic_sessions()
}
}
