% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{enable_tab_completion}
\alias{enable_tab_completion}
\title{Enable tab completion in the readline REPL}
\usage{
enable_tab_completion()
}
\description{
Enable tab completion in the readline REPL
}
\keyword{internal}
