% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{execute_tool}
\alias{execute_tool}
\title{Execute a tool call and return the result}
\usage{
execute_tool(tool_name, arguments)
}
\arguments{
\item{tool_name}{Name of the tool to execute}

\item{arguments}{Named list of arguments}
}
\value{
Character string with the result
}
\description{
Execute a tool call and return the result
}
\keyword{internal}
