% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{extract_r_code_blocks}
\alias{extract_r_code_blocks}
\title{Extract R code blocks from markdown text}
\usage{
extract_r_code_blocks(text)
}
\description{
Supports \verb{```{r}...```} and \verb{```...```} blocks
}
\keyword{internal}
