% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{load_r_history}
\alias{load_r_history}
\title{Write a turn to the session history file}
\usage{
load_r_history()
}
\description{
Write a turn to the session history file
}
\keyword{Load}
\keyword{R}
\keyword{command}
\keyword{for}
\keyword{history}
\keyword{internal}
\keyword{recall}
\keyword{up-arrow}
