% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{mcp_all_tools}
\alias{mcp_all_tools}
\title{Get all MCP tool definitions (merged from all connected servers)}
\usage{
mcp_all_tools()
}
\description{
Get all MCP tool definitions (merged from all connected servers)
}
\keyword{internal}
