% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{mcp_send_notification}
\alias{mcp_send_notification}
\title{Send a JSON-RPC notification (no response expected)}
\usage{
mcp_send_notification(srv, method, params)
}
\description{
Send a JSON-RPC notification (no response expected)
}
\keyword{internal}
