% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{mcp_tools}
\alias{mcp_tools}
\title{Get OpenAI-compatible tool definitions from an MCP server}
\usage{
mcp_tools(srv)
}
\description{
Get OpenAI-compatible tool definitions from an MCP server
}
\keyword{internal}
