% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.agenticr_config}
\alias{print.agenticr_config}
\title{Print current configuration}
\usage{
\method{print}{agenticr_config}(x, ...)
}
\arguments{
\item{x}{An agenticr_config object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print current configuration
}
