% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{process_input}
\alias{process_input}
\title{Process user input: detect NL vs R and route accordingly}
\usage{
process_input(input)
}
\description{
Process user input: detect NL vs R and route accordingly
}
\keyword{internal}
