% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{PROVIDER_PRESETS}
\alias{PROVIDER_PRESETS}
\title{Provider presets for common LLM APIs}
\usage{
PROVIDER_PRESETS
}
\description{
Use agentic_config(provider = "name") to switch providers.
}
\keyword{internal}
