% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{remove_r_code_blocks}
\alias{remove_r_code_blocks}
\title{Remove R code blocks from text, leaving only explanatory text}
\usage{
remove_r_code_blocks(text)
}
\description{
Remove R code blocks from text, leaving only explanatory text
}
\keyword{internal}
