% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tool_grep_search}
\alias{tool_grep_search}
\title{Search file contents for a regex pattern using rg or grep}
\usage{
tool_grep_search(pattern, path = ".", context_lines = 2)
}
\description{
Search file contents for a regex pattern using rg or grep
}
\keyword{internal}
