% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tool_search_variables}
\alias{tool_search_variables}
\title{Search for variables in the environment}
\usage{
tool_search_variables(pattern)
}
\description{
Search for variables in the environment
}
\keyword{internal}
