\name{print.agenticr_config}
\alias{print.agenticr_config}
\title{Print agenticr configuration}
\description{Print agenticr configuration}
\usage{
\method{print}{agenticr_config}(x, ...)
}
\arguments{
\item{x}{An agenticr_config object}
\item{...}{Additional arguments (ignored)}
}
\examples{
\dontrun{
agentic_config()
}
}
